backend/requirements.txt
fastapi
uvicorn
pydantic
google-adk
google-genai
python-dotenv
supabase
psycopg2-binary
pyjwt
litellm